pyproject.toml
[project]
name = "kalshi-autotrader"
version = "0.1.0"
description = "Autonomous Kalshi trading system scaffold with LLM research pipeline"
readme = "README.md"
authors = [{name = "Your Team", email = "trading@example.com"}]
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.26.0",
    "websockets>=12.0",
    "pydantic>=2.8.0",
    "pyyaml>=6.0",
    "rich>=13.7.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "tenacity>=8.2.0",
    "cryptography>=43.0.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.5.0",
    "pandas>=2.2.0",
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.0",
    "prometheus-client>=0.20.0",
    "python-json-logger>=2.0.7"
]

[project.optional-dependencies]
research = [
    "openai>=1.40.0",
    "langchain>=0.2.0"
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
    "testcontainers[postgresql]>=4.5.0"
]
secrets = [
    "hvac>=2.2.0"
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
